Create Physical_class_tree
This program has been copied from Chapter 29 of Learning Python, 5th Edition. I have copied here for future utilization, as this function can be used on any module containing a class hierarchy. It will result in a nice physical display of class trees and heights when run in a GUI display. May be useful to trace classes in programs. Run self test in a GUI to see example.