Create FET_FST_combiner.py
Should be final version of this script.  Works fine, but there is an artifact in the CSV file that could be removed (parenthesis in the two columns of the key tuple). Can't seem to find a way to remove these without errors in the script, but they don't matter much outside of aesthetic appeal of the CSV file.  Won't matter for my personal analysis, but if file wanted more global use in the future is a problem I might want to address.